requirements.txt
PyQt6
grpcio
grpcio-tools
